Recreate C# bindings as a dotnet solution, rename List<T> -> MapperList<T>
