+++ bindings/csharp/Mapper.NET/Time.cs
@@ -1,0 +1,128 @@
+using System.Runtime.InteropServices;
+
+namespace Mapper.NET;
+
+public class Time
+    {
+        // internal long _time;
+
+        [StructLayout(LayoutKind.Explicit)]
+        internal struct timeStruct
+        {
+            [FieldOffset(0)]
+            internal long ntp;
+            [FieldOffset(0)]
+            internal UInt32 sec;
+            [FieldOffset(4)]
+            internal UInt32 frac;
+        }
+        internal timeStruct data;
+
+        public Time(long ntp)
+            { data.ntp = ntp; }
+        public Time(Time time)
+            { data.ntp = time.data.ntp; }
+        public Time()
+            { data.sec = 0; data.frac = 1; }
+        public Time(double seconds)
+            { this.SetDouble(seconds); }
+
+        [DllImport("mapper", CharSet = CharSet.Ansi, CallingConvention = CallingConvention.StdCall)]
+        private static extern unsafe int mpr_time_set(IntPtr l, long r);
+        public unsafe Time Set(Time time)
+        {
+            fixed (void* t = &data)
+            {
+                mpr_time_set((IntPtr)t, time.data.ntp);
+            }
+            return this;
+        }
+
+        public Time SetDouble(Double seconds)
+        {
+            if (seconds > 0.0)
+            {
+                data.sec = (UInt32)Math.Floor(seconds);
+                seconds -= data.sec;
+                data.frac = (UInt32) (((double)seconds) * 4294967296.0);
+            }
+            else
+                data.ntp = 0;
+            return this;
+        }
+
+        private static double as_dbl(Time time)
+        {
+            return (double)time.data.sec + (double)time.data.frac * 0.00000000023283064365;
+        }
+
+        public Time Add(Time addend)
+        {
+            data.sec += addend.data.sec;
+            data.frac += addend.data.frac;
+            if (data.frac < addend.data.frac) /* overflow */
+                ++data.sec;
+            return this;
+        }
+
+        public Time Subtract(Time subtrahend)
+        {
+            if (data.sec > subtrahend.data.sec)
+            {
+                data.sec -= subtrahend.data.sec;
+                if (data.frac < subtrahend.data.frac) /* overflow */
+                    --data.sec;
+                data.frac -= subtrahend.data.frac;
+            }
+            else
+                data.ntp = 0;
+            return this;
+        }
+
+        public Time Multiply(double multiplicand)
+        {
+            if (multiplicand > 0.0)
+            {
+                multiplicand *= Time.as_dbl(this);
+                data.sec = (UInt32) Math.Floor(multiplicand);
+                multiplicand -= data.sec;
+                data.frac = (UInt32) (multiplicand * 4294967296.0);
+            }
+            else
+                data.ntp = 0;
+            return this;
+        }
+
+        /* casting between Time and double */
+        public static implicit operator double(Time t) => Time.as_dbl(t);
+        public static explicit operator Time(double d) => new Time(d);
+
+        /* Overload some arithmetic operators */
+        public static Time operator +(Time a, Time b) => new Time(a).Add(b);
+
+        public static Time operator -(Time a, Time b) => new Time(a).Subtract(b);
+
+        public static Time operator *(Time a, double b) => new Time(a).Multiply(b);
+
+        public static bool operator ==(Time a, Time b) => a.data.ntp == b.data.ntp;
+        public static bool operator !=(Time a, Time b) => a.data.ntp != b.data.ntp;
+        public static bool operator >(Time a, Time b) => a.data.ntp > b.data.ntp;
+        public static bool operator <(Time a, Time b) => a.data.ntp < b.data.ntp;
+        public static bool operator >=(Time a, Time b) => a.data.ntp >= b.data.ntp;
+        public static bool operator <=(Time a, Time b) => a.data.ntp <= b.data.ntp;
+
+        public override bool Equals(object o)
+        {
+            if (o == null)
+                return false;
+            var second = o as Time;
+            return second != null && data.ntp == second.data.ntp;
+        }
+
+        public override int GetHashCode()
+        {
+            return (int)(data.sec ^ data.frac);
+        }
+
+        public override string ToString() => $"Mapper.Time:{this.data.sec}:{this.data.frac}";
+    }
